1. update test snippet; 2. update vscode debug;

+++ snippet/src/memory.cpp
@@ -3,49 +3,53 @@
 
 using namespace std;
 
-class Parent {
-    public:
-        int iparent;
-        Parent ():iparent (10) {}
-        virtual void f() { cout << " Parent::f()" << endl; }
-        virtual void g() { cout << " Parent::g()" << endl; }
-        virtual void h() { cout << " Parent::h()" << endl; }
+class Parent
+{
+public:
+    int iparent;
+    Parent(): iparent(10) {}
+    virtual void f() { cout << " Parent::f()" << endl; }
+    virtual void g() { cout << " Parent::g()" << endl; }
+    virtual void h() { cout << " Parent::h()" << endl; }
 };
 
 
 
-class Child : public Parent {
-    public:
-        int ichild;
-        Child():ichild(100) {}
-        virtual void f() { cout << "Child::f()" << endl; }
-        virtual void g_child() { cout << "Child::g_child()" << endl; }
-        virtual void h_child() { cout << "Child::h_child()" << endl; }
+class Child : public Parent
+{
+public:
+    int ichild;
+    Child(): ichild(100) {}
+    virtual void f() { cout << "Child::f()" << endl; }
+    virtual void g_child() { cout << "Child::g_child()" << endl; }
+    virtual void h_child() { cout << "Child::h_child()" << endl; }
 };
 
 
 
-class GrandChild : public Child{
-    public:
-        int igrandchild;
-        GrandChild():igrandchild(1000) {}
-        virtual void f() { cout << "GrandChild::f()" << endl; }
-        virtual void g_child() { cout << "GrandChild::g_child()" << endl; }
-        virtual void h_grandchild() { cout << "GrandChild::h_grandchild()" << endl; }
+class GrandChild : public Child
+{
+public:
+    int igrandchild;
+    GrandChild(): igrandchild(1000) {}
+    virtual void f() { cout << "GrandChild::f()" << endl; }
+    virtual void g_child() { cout << "GrandChild::g_child()" << endl; }
+    virtual void h_grandchild() { cout << "GrandChild::h_grandchild()" << endl; }
 };
 
-int main(){
+int main()
+{
 
     typedef void(*Fun)(void);
     GrandChild gc;
     Fun pFun;
 
-    int** pVtab = (int**)&gc;
+    int64_t ** pVtab = (int64_t **)&gc;
 
     cout << "[0] GrandChild::_vptr->" << endl;
-    for (int i=0; (Fun)pVtab[0][i]!=NULL; i++){
+    for(int i = 0; (Fun)pVtab[0][i] != NULL; i++) {
         pFun = (Fun)pVtab[0][i];
-        cout << "    ["<<i<<"] ";
+        cout << "    [" << i << "] ";
         pFun();
 
     }
